Implement getenv for non-Windows OSes
As Windows does not have a populated environ of U8s, it'll take a slightly different solution. I did some looking at the Zig stdlib os.getenv, and it uses the 0 byte to find the end of the environment variable, along with environ being a slice of [:0]const u8, so I have no clue why it isn't returned as [:0]const u8. Might be an issue in os.getenv